+++ src/docs/components/construction.svelte
@@ -4,7 +4,7 @@
 </script>
 
 <div
-	class="my-4 flex flex-col items-start gap-2 rounded-md border border-magnum-500 bg-magnum-500/10 p-4
+	class="my-4 flex flex-col items-start gap-2 rounded-xl border border-magnum-500 bg-magnum-500/10 p-4
 	lg:flex-row lg:items-center"
 >
 	<Construction class="shrink-0 text-magnum-300 square-8 lg:square-6" />