+++ src/docs/components/preview-wrapper.svelte
@@ -3,7 +3,7 @@
 </script>
 
 <div class="comp-preview h-[20rem] overflow-x-auto lg:h-[28rem] lg:overflow-x-hidden">
-	<div class={cn('mx-auto inline-block px-4')}>
+	<div class={cn('relative z-10 mx-auto inline-block px-4')}>
 		<slot />
 	</div>
 </div>
@@ -11,16 +11,28 @@
 <style lang="postcss">
 	.comp-preview {
 		display: flex;
-		/* justify-content: center; */
 		align-items: center;
 
-		background: linear-gradient(315deg, theme('colors.magnum.600'), theme('colors.magnum.700'));
-
-		border-radius: theme('borderRadius.lg');
-		/* padding: theme('spacing.8'); */
+		background-color: theme('colors.magnum.400');
+		border-radius: theme('borderRadius.xl');
+		position: relative;
 
 		@media screen('md') {
 			padding: theme('spacing.12');
+		}
+
+		&::before {
+			position: absolute;
+			content: '';
+			inset: 0;
+			background-image: radial-gradient(
+				circle at 1px 1px,
+				theme('colors.magnum.900/0.25') 1px,
+				transparent 0
+			);
+			background-size: 1rem 1rem;
+			background-repeat: repeat;
+			background-position: 0.5rem center;
 		}
 	}
 </style>