+++ src/routes/(landing-ui)/accordion.svelte
@@ -34,10 +34,7 @@
 </script>
 
 <div
-	class={cn(
-		'mx-auto w-[18rem] max-w-full rounded-xl bg-white shadow-lg sm:w-[25rem]',
-		className,
-	)}
+	class={cn('mx-auto w-[18rem] max-w-full rounded-xl bg-white shadow-lg sm:w-[25rem]', className)}
 	{...$root}
 >
 	{#each items as { id, title, description }, i}
@@ -54,7 +51,7 @@
 						'bg-white px-5 py-5 text-base font-medium leading-none',
 						'text-black transition-colors hover:bg-neutral-100 focus:!ring-0',
 						'focus-visible:text-magnum-800',
-						i !== 0 && 'border-t border-t-neutral-300',
+						i !== 0 && 'border-t border-t-neutral-300'
 					)}
 				>
 					{title}
@@ -62,10 +59,7 @@
 			</h2>
 			{#if $isSelected(id)}
 				<div
-					class={cn(
-						'content',
-						'overflow-hidden bg-neutral-100 text-sm text-neutral-600',
-					)}
+					class={cn('content', 'overflow-hidden bg-neutral-100 text-sm text-neutral-600')}
 					use:melt={$content(id)}
 					transition:slide
 				>