+++ src/routes/+layout.svelte
@@ -1,4 +1,14 @@
 <script lang="ts">
+	import '@fontsource/inter/100.css';
+	import '@fontsource/inter/200.css';
+	import '@fontsource/inter/300.css';
+	import '@fontsource/inter/400.css';
+	import '@fontsource/inter/500.css';
+	import '@fontsource/inter/600.css';
+	import '@fontsource/inter/700.css';
+	import '@fontsource/inter/800.css';
+	import '@fontsource/inter/900.css';
+	import '../fonts.css';
 	import '../app.postcss';
 	import { dev } from '$app/environment';
 	import { JsIndicator, SiteHeader, TailwindIndicator } from '$docs/components/index.js';